% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghstudio.R
\name{view_diff}
\alias{view_diff}
\title{view diff between commits}
\usage{
view_diff(
  sha1 = "HEAD",
  sha2 = paste0(sha1, "~"),
  mode = c("auto", "unified", "sidebyside", "context"),
  context = 2,
  ...
)
}
\arguments{
\item{sha1, sha2}{sha or branch name}

\item{mode}{character(1L), one of:
\itemize{
  \item \dQuote{unified}: diff mode used by \code{git diff}
  \item \dQuote{sidebyside}: line up the differences side by side
  \item \dQuote{context}: show the target and current hunks in their
    entirety; this mode takes up a lot of screen space but makes it easier
    to see what the objects actually look like
  \item \dQuote{auto}: default mode; pick one of the above, will favor
    \dQuote{sidebyside} unless \code{getOption("width")} is less than 80,
    or in \code{diffPrint} and objects are dimensioned and do not fit side
    by side, or in \code{diffChr}, \code{diffDeparse}, \code{diffFile} and
    output does not fit in side by side without wrapping
}}

\item{context}{integer(1L) how many lines of context are shown on either side
of differences (defaults to 2).  Set to \code{-1L} to allow as many as
there are.  Set to \dQuote{auto}  to display as many as 10 lines or as few
as 1 depending on whether total screen lines fit within the number of lines
specified in \code{line.limit}.  Alternatively pass the return value of
\code{\link[diffobj]{auto_context}} to fine tune the parameters of the auto context
calculation.}

\item{...}{Passed to diffobj::diffFile}
}
\value{
Returns \code{NULL} invisibly, called for side effects.
}
\description{
view diff between commits
}
