% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghstudio.R
\name{view_issues}
\alias{view_issues}
\alias{view_assigned}
\alias{view_created}
\alias{view_mentions}
\title{View issues}
\usage{
view_issues(
  author = NULL,
  assignee = NULL,
  mentions = NULL,
  open = TRUE,
  pr = NA,
  repo = current_repo()
)

view_assigned(
  user = github_name(),
  open = TRUE,
  pr = NA,
  repo = current_repo()
)

view_created(user = github_name(), open = TRUE, pr = NA, repo = current_repo())

view_mentions(
  user = github_name(),
  open = TRUE,
  pr = NA,
  repo = current_repo()
)
}
\arguments{
\item{author}{A string or \code{NULL}}

\item{assignee}{A string or \code{NULL}}

\item{mentions}{A string or \code{NULL}}

\item{open}{A boolean, not \code{NA}}

\item{pr}{A boolean, \code{NA} for both PRs and issues}

\item{repo}{A string}

\item{user}{A string}
}
\value{
Returns \code{NULL} invisibly, called for side effects.
}
\description{
View issues in viewer, \code{view_assigned()}, \code{view_created()} and \code{view_mentions()}
are wrappers around \code{view_issues()}.
}
